pyproject.toml
[project]
name = "alpr"
version = "2.0.0"
description = "Automatic License Plate Recognition - ALPR 2.0 com pipeline local PaddleOCR e validacao Premium opcional via Plate Recognizer"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["alpr", "ocr", "license-plate-recognition", "computer-vision", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "opencv-contrib-python>=4.7.0",
    "numpy>=1.24.0",
    "streamlit>=1.52.0",
    "ultralytics>=8.0.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "pillow>=9.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.60.0",
    "matplotlib>=3.4.0",
    "blinker>=1.9.0",
    "typing-extensions>=4.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "httpx>=0.24.0",
    "av>=10.0.0",
    "moviepy>=1.0.3",
    "paddleocr>=3.4.1",
    "paddlepaddle>=3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
    "respx>=0.20.0",
    "pre-commit>=3.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
# Por padrão NÃO rodar testes marcados como 'integration'.
# Para rodá-los explicitamente:
#   pytest -m integration
addopts = "-m 'not integration'"
markers = [
    "integration: testes que dependem de recursos externos ou mídia real (lentos, ~10s+)",
]

# =====================================================================
# Black — formatador de código
# =====================================================================
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    venv
  | \.venv
  | build
  | dist
  | models
  | data
)/
'''

# =====================================================================
# Ruff — linter rápido (substitui flake8, isort, pyupgrade, etc.)
# =====================================================================
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
    "models",
    "data",
]

[tool.ruff.lint]
# Regras ativas:
#   E, W  → pycodestyle
#   F     → pyflakes (unused imports, undefined names)
#   I     → isort (ordenação de imports)
#   UP    → pyupgrade (sintaxe moderna)
#   B     → flake8-bugbear (bugs comuns)
#   SIM   → flake8-simplify
#   RUF   → regras próprias do ruff
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
# Ignorar:
#   E501 (line too long — já coberto por black com tolerância)
#   B008 (function call in default arg — padrão comum em typer/fastapi)
#   SIM108 (ternary — às vezes menos legível)
#   RUF001/RUF002/RUF003 (caracteres unicode ambíguos — acentos PT-BR são legítimos)
ignore = ["E501", "B008", "SIM108", "RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.per-file-ignores]
# Testes podem ter asserts, imports não usados (fixtures), mutable defaults em
# fixtures stub, etc.
"tests/**/*.py" = ["F401", "F811", "B011", "RUF012"]
# __init__.py pode re-exportar sem usar
"**/__init__.py" = ["F401"]
# conftest.py manipula sys.path antes de importar src.* — E402 é intencional
"tests/conftest.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# =====================================================================
# Mypy — type checker (modo permissivo inicial)
# =====================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
# Modo permissivo — strict pode ser ativado gradualmente por módulo
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
show_error_codes = true
exclude = [
    "venv/",
    ".venv/",
    "build/",
    "dist/",
    "models/",
    "data/",
    "tests/",
]

# Módulos externos sem stubs — ignorar silenciosamente
[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "ultralytics.*",
    "paddleocr.*",
    "skimage.*",
    "scipy.*",
    "sklearn.*",
    "moviepy.*",
    "av.*",
]
ignore_missing_imports = true
